move worker to root
